+++ changelog.md
@@ -3,6 +3,7 @@
 - Vintage Comb Block is no longer missing an in-world model
 - Escritoire Game state now persists across world reloads
 - Add new Forestry: CE logo
+- Add Sengir's Notes to the creative menu
 
 ## Forestry 2.10.2
 - Fix bugs in French translation, thanks to Horusbk (#313)